+++ packages/react-native-callingx/android/src/main/java/io/getstream/rn/callingx/CallService.kt
@@ -25,6 +25,7 @@ import io.getstream.rn.callingx.notifications.NotificationChannelsManager
 import io.getstream.rn.callingx.notifications.NotificationsConfig
 import io.getstream.rn.callingx.repo.CallRepository
 import io.getstream.rn.callingx.repo.CallRepositoryFactory
+import io.getstream.rn.callingx.utils.AudioEndpointUtils
 import io.getstream.rn.callingx.utils.LifecycleListener
 import io.getstream.rn.callingx.utils.SettingsStore
 import kotlinx.coroutines.CancellationException
@@ -357,6 +358,8 @@ class CallService : Service(), CallRepository.Listener {
                 }
             }
             is Call.None, is Call.Unregistered -> {
+                CallRegistrationStore.removeTrackedCall(callId)
+                AudioEndpointStore.clear(callId)
                 repromoteForegroundIfNeeded(callId)
                 if (!callRepository.hasRingingCall()) notificationManager.stopRingtone()
 
@@ -428,10 +431,18 @@ class CallService : Service(), CallRepository.Listener {
         }
     }
 
-    override fun onCallEndpointChanged(callId: String, endpoint: String) {
-        sendBroadcastEvent(CallingxModuleImpl.CALL_ENDPOINT_CHANGED_ACTION) {
+    override fun onCallAudioEndpointsChanged(callId: String) {
+        val call = callRepository.getCall(callId) ?: return
+        val snapshotJson =
+                AudioEndpointUtils.snapshotJson(
+                        call.currentCallEndpoint,
+                        call.availableCallEndpoints,
+                )
+        AudioEndpointStore.setSnapshot(callId, snapshotJson)
+
+        sendBroadcastEvent(CallingxModuleImpl.CALL_AUDIO_ENDPOINTS_CHANGED_ACTION) {
             putExtra(CallingxModuleImpl.EXTRA_CALL_ID, callId)
-            putExtra(CallingxModuleImpl.EXTRA_AUDIO_ENDPOINT, endpoint)
+            putExtra(CallingxModuleImpl.EXTRA_AUDIO_ENDPOINTS_SNAPSHOT, snapshotJson)
         }
     }